Add files via upload
This is the evolution of the initial development done around the ADXL345 sensor.
The ADXL345 is great but its resolution is quite poor. Instead of spending time in averaging its output to get some usable output I searched for a better sensor in a similar price range. The MMA8451 turns out to be the right product to use. It works on 14bits where the ADXL345 only has 10bits. 
Have fun !